version 1.1.6.
Added prelude, making RTile !Send and !Sync

As RTile are in general stored in thread locals to be able to combine multiple RTiles for code generation, thought it would be better to make them !Send and !Sync.
So, making the struct RTile !Send and !Sync by adding PhantomData<Rc<()>> to RTile.

remove rustfmt nightly build options
updated the readme.md with a shorter example